+++ actix-session/src/storage/mod.rs
@@ -17,7 +17,14 @@ mod redis_actor;
 #[cfg(feature = "redis-rs-session")]
 mod redis_rs;
 
-#[cfg(any(feature = "redis-actor-session", feature = "redis-rs-session"))]
+#[cfg(feature = "sqlite-session")]
+mod sqlite;
+
+#[cfg(any(
+    feature = "redis-actor-session",
+    feature = "redis-rs-session",
+    feature = "sqlite-session"
+))]
 mod utils;
 
 #[cfg(feature = "cookie-session")]
@@ -26,3 +33,5 @@ pub use cookie::CookieSessionStore;
 pub use redis_actor::{RedisActorSessionStore, RedisActorSessionStoreBuilder};
 #[cfg(feature = "redis-rs-session")]
 pub use redis_rs::{RedisSessionStore, RedisSessionStoreBuilder};
+#[cfg(feature = "sqlite-session")]
+pub use sqlite::SqliteSessionStore;